Fixed bug in repost detector
I suppose the Discord API now uses discord.com links for message jump URLs instead of discordapp.com. The bug caused malformed information to be added to the database, which caused errors every time the bot detected reposted content and attempted to fetch the jump URL to the original message from the database.