Update Docker configuration
Files have been moved into their respective directories so they can be mounted better within the Docker container.

Environment variables for pip commands have been moved to environment variable arguments in the docker run command.

The method the current commit ref is fetched has been adjusted as the .git directory is not mounted within the container, instead just the master ref file is mounted in the container and that is now read from.