Scrap Python bot
discord.py has been deprecated for ages, but I never got around to updating the bot and moving to another library. However, on my private Git server I was working on a Discord bot in Node.js from scratch (i.e. without libraries such as discord.js) which I really enjoyed doing. Though that bot has not been updated to the new Discord API in a while, plus I attempted to move it to TypeScript and broke more or less all of the code :/

My plan is to slowly rewrite my 'Discord Bot from Scratch' in Node.js using TypeScript on this repository, and update it to work with the latest Discord API changes.